requirements.txt
strands-agents
strands-agents-tools
google-generativeai
pydantic
python-dotenv
strands-agents[gemini]